NP-2155: Unit tests for CristinApiClient into its own test class
